requirements.txt
python-telegram-bot==22.0
python-dotenv==1.0.0
loguru==0.7.2
